Added back button support by using a dialog, converted to using a library.
