+++ lib/src/enum.dart
@@ -1,0 +1,9 @@
+part of direct_select_plugin;
+
+enum DirectSelectMode {
+  /// The [DirectSelect] is engaged by dragging on it.
+  drag,
+
+  /// The [DirectSelect] is engaged by tapping on it.
+  tap,
+}
